CHORE: move event enums to kafka module
