+++ notification/core/src/main/kotlin/vp/togedo/data/sse/SSEDao.kt
@@ -1,5 +1,7 @@
 package vp.togedo.data.sse
 
+import vp.togedo.kafka.data.enums.EventEnums
+
 data class SSEDao (
     val event: EventEnums,
     val sender: String,
